the pixel XYZCT matrix is now explicitely cast to matlab type uint8.  The type of the matlab matrix is very important because it is used to set the bitsPerPixel for OMEIS's newPIxels call.
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@4713 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/matlab/ome_pixelsTest.m
@@ -35,4 +35,4 @@
 % test the Matlab handler handling of the pixels semantic type
 function [outPixels] = ome_pixelsTest (inPixels)
 
-outPixels = inPixels';
+outPixels = uint8(inPixels');
